update sample app and Docker files
