+++ app/.gitignore
@@ -10,6 +10,7 @@
 
 # next.js
 /.next/
+/_next/
 /out/
 
 # production
