+++ app/next.config.js
@@ -1,0 +1,5 @@
+// /** @type {import('next').NextConfig} */
+module.exports = {
+  reactStrictMode: true,
+  distDir: '_next',
+};
