+++ app/styles/Home.module.css
@@ -9,12 +9,28 @@
 }
 
 .main {
-  padding: 5rem 0;
+  padding: 7rem 0;
   flex: 1;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  padding-top: 50px;
+  overflow: auto;
+  text-align: center;
+}
+
+.homeHero {
+  background: url("/home.webp");
+  background-size: cover;
+  background-position: center center;
+  width: 100%;
+  height: 80vh;
+}
+
+.homeHero h1 {
+  margin: 30% 0px;
+  line-height: 1.5;
 }
 
 .footer {
@@ -70,6 +86,7 @@
 }
 
 .grid {
+  cursor: pointer;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -95,6 +112,7 @@
 .card:active {
   color: #0070f3;
   border-color: #0070f3;
+  cursor: pointer;
 }
 
 .card h2 {
