+++ package.json
@@ -1,5 +1,5 @@
 {
   "dependencies": {
-    "create-next-app": "^11.0.1"
+    "create-next-app": "^11.1.4"
   }
 }
