--- sample-app/next-env.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="next" />
-/// <reference types="next/types/global" />
-/// <reference types="next/image-types/global" />
