--- sample-app/next.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  reactStrictMode: true,
-}
